Refactor: move validation methods into ValidationHelper module
